refactored into branchLogo component
